Docs: add the "How To - Finding and querying for data" section (#4080)
Move in the sections on the `QueryBuilder` to the How-to section called

    "Finding and querying for data"

The original text was too extensive for the purposes of the How-to and
so it has been split in two. The detailed information on advanced
queries will be added in another PR in the section "Topic - Database".

The text for this How-to has been adapted from the original text with
significant changes in the structure and simplifying the examples. Here
wherever possible they use the `ArithmeticAddCalculation` for consistency
with other sections and tutorials. The new structure is as follows:

    * Selecting entities
    * Retrieving results
    * Filters
    * Relationships
    * Projections