Add tests for new `MultiplyAddWorkChain` and `add_multiply` plugins
Add also tests for the `ArithmeticAddCalculation` which is not new but
was up till now untested. Now that this will start to be used heavily in
tutorials and documentation for teaching purposes, we should make sure
that it is well tested. Here we add an integration test and a unit test
for the `prepare_for_submission` method.