Enable loading config.yml files from URL for ceratin `verdi` commands (#3977)
In the view of the plans of starting online repositories that maintain
yaml configuration files for computers and codes, it becomes very useful
to be able to directly import a configuration file from a URL.

Version 0.6 of the `config-file-option` now respects the `type` parameter,
and file handles can be passed directly to the configuration file
provider. This allows handling opening files locally and from URLs on
the same footing.