`QueryBuilder`: fix type string filter generation for `Group` subclasses (#4144)
The `aiida.orm.querybuilder.get_group_type_filter` function that is
responsible for generating the correct filters for the `type_string`
when a `Group` or subclass of it is appended, contained a bug. To
generate the correct type string, it should strip the `group.` prefix
from the provided ORM classifier. This was incorrectly done with the
string method `lstrip`, which does not strip an entire prefix as is, but
will strip any characters that are matched starting from the left. This
would cause extra characters to be erroneously stripped if the rest of
the type string started with any letters in the sequence `group.`. For
example `group.pseudo.family` would be stripped to `seudo.family`. The
solution is to use the first N characters where N is the length of the
prefix.