Wrap communicator of `Runner` in `plumpy.LoopCommunicator`
This is necessary because otherwise subscribers that use the runner's
communicator and receive a message will execute the callback on the
communication thread. This should be prevented for two reasons. Firstly,
it will cause the business logic to be executed on the communication
thread which can cause it to fail to respond in time to heartbeat
requests from RabbitMQ when it is under heavy load. Secondly, the
communication thread does not share the same session as the main thread.
This can cause problems when the callbacks query the database as it can
be out of sync with the session of the main thread. For Django this will
most likely go by unnoticed, but SqlAlchemy will except.

Since the communicator of the runner is now wrapped, the process
launcher of the daemon runner no longer has to be added as a task
subscriber through a callback, but the `add_task_subscriber` can be
called directly and the `LoopCommunicator` will take of everything.