Fix the unhandled failure mechanism of the `BaseRestartWorkChain`
Any `BaseRestartWorkChain` will fail if the sub process fails twice in a
row without being handled by a registered process handler. To monitor
this the `unhandled_failure` context variable is used, which is set to
`True` once a failed process was not handled. It should be unset as soon
as the next process finishes successfully, or the failed process is
handled. The reset was there, but was incorrectly setting the context
var to `True` instead of `False`, which went unnoticed due to a lack of
tests. Tests are now added, as well as for the case where the sub
process excepted or was killed.