Simplify `ArithmeticAddCalculation/ArithmeticAddParser` implementation
The purpose of these basic `CalcJob` and `Parser` plugins is for easy
integration testing, but also to serve as an example for the
documentation and tutorials on how to implement these plugin types.

To prevent code duplication and make sure that the code that is
presented in the documentation is actually tested, we want to use this
actual implementation there by literal-including it. This means that the
code has to be as simple and minimal as is possible, within reason and
by maintaining correctness for testing purposes.

That is why we get rid of unnecessary utility methods and try to
implement the code as straightforward as possible. Since the code is
simplified it fits well into a single method anyway.