Update tutorial and how-to guides to literal-include for code snippets
The basic tutorial and how-to guides on implementing `CalcJob` and
`Parser` plugins, as well as how-to write work functions and work chains
have been updated to `literal-include` the implementation of the actual
plugins that ship with `aiida-core`.

This has the advantage that the code that is shown in these crucial
parts of the documentation is guaranteed to work with each release as
the plugins themselves are tested in the unit test suite. The down side,
however, is that we do not always want to show everything from the
entire implementation at any point, so we have to resort to specifying
line numbers.

This is fine, except that it runs the risk that the line numbers run out
of sync if the implementation is changed and automatically detecting
this is difficult. Nonetheless, the other solution of just duplicating
the code in the documentation has the same problem, and in addition the
code is not easily tested. Tests *could* be added, but given that
testing these plugins is not trivial, the required additional code and
infrastructure would be more of a burden than the current solution.

Finally, by using the `:pyobject:`, `:start-after:` and `:end-before:`
options as much as possible, depending on explicit line numbers is
also reduced. Currently, only a snippet in `howto/codes.rst` on exit
codes in parsers relies on explicit line numbers. That is why the
implementation in `aiida.parsers.plugins.arithmetic.add` has a comment
with a warning that if edited, the snippet in the documentation should
be checked for consistency.