Add parameter `custom_cell` to `TrajectoryData.get_step_structure`
If a `TrajectoryData` does not contain cells, the method `get_step_structure`
returns a `StructureData` with the default cell.
The optional parameter `custom_cell` allows one to specify the cell of
the structure that will be extracted from a `TrajectoryData`.
If omitted, the cell will be read from the trajectory, if present.
If not found, the `StructureData` will use the default cell, and an user
warning will be emitted.
Docs updated.