`SshTransport.gettree`: allow non-existing nested target directories
The `gettree` method would raise an `OSError` for the `SshTransport` if
the target local path contains intermediate subdirectories that are not
created beforehand. The same would work without problem for the
`LocalTransport` plugin. It makes sense to just create non-existing
directories along the way, so the `os.mkdir` command that was used
originally to create the target directory is replaced with the method
`os.makedirs` setting `exists_ok=True`.