`WorkChain`: guarantee to maintain order of appended awaitables (#4156)
Intuitively, one would expect the order of awaitables in the context
once they have been resolved, to maintain the same order with which they
had been inserted using the `append_` construct in the `to_context`
call. However, this was not the case, as the resolved awaitables were
just appended to the target container in the order that they resolved.
To correct this behavior, once the awaitable is inserted, it is already
inserted in the context, where its resolved value should be placed in
the future. This allows the `resolve_awaitable` method to find the
correct place and simply swap out the awaitable, that has been serving
as a placeholder, with the resolved value.