Add support for TLS/SSL in message broker configuration
Add two new keys to a profile configuration:

 * broker_virtual_host
 * broker_parameters

The `broker_virtual_host, by default `/` can be changed to use a
different virtual host. Note that it has to start with a forward slash.

If SSL is required, the `broker_protocol` should be set to `amqps`. The
`broker_parameters` can be used to specify the specific SSL parameters.
The following parameters can be set:

 * heartbeat  # heartbeat timeout in seconds
 * cafile  # string containing path to ca certificate file
 * capath  # string containing path to ca certificates
 * cadata  # base64 encoded ca certificate data
 * keyfile  # string containing path to key file
 * certfile  # string containing path to certificate file
 * no_verify_ssl  # boolean disables certificates validation should be "0" or "1"

Currently it is not clear how the version of TLS/SSL can be specified.