Restore deleted images for `README.md` and update acknowledgements
Commit `94a40ae4768e7bcefce8409bed50ec9487b0e4c2` accidentally removed
the AiiDA logo because it was replaced with a logo specifically designed
for the docs. However, the logo was also being used by the `README.md`.
The logo is restored and renamed for consistency. In addition, the
acknowledgments are simplified as they are in the docs, by removing the
images of logo's of the supporting agencies and referring to the AiiDA
website for a complete list.