`verdi status`: add `--no-rmq` flag to skip the RabbitMQ check (#4181)
This makes it possible to use the command to check whether the system is
up to at least query the database, for which RabbitMQ is not necessary.
The behavior of `verdi status` is also adapted to no longer return a
non-zero exit code if the daemon is not running. The daemon not running
is not the same as not being able to connect to the Postgres database or
the RabbitMQ message broker. So it shouldn't return a non-zero exit code
either. If the user wants to poll this, the command `verdi daemon status`
should be used instead.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>