`CalcJobResultManager`: fix bug that broke tab completion
For tab-completion to work properly, the `__getattr__` method need to
properly return an `AttributeError` if the corresponding attribute does
not exist, however, it was accidentally raising a `KeyError` since the
implementation caught an `AttributeError` but the dereferencing was
using `__getitem__` so would never raise an `AttributeError`.

Likewise, the `__getitem__` implementation was incorrectly returning an
`AttributeError` instead of a `KeyError` making it behave different from
any other mapping-like data structure.