`NodeTranslator`: do not assume `get_export_formats` exists
The `get_all_download_formats` will try to call `get_export_formats` on
all the known plugin types. However, it is not guaranteed that all
plugins implement this method. This was properly considered in the case
that `full_type` is not None, however, in the other case the
`AttributeError` was not being caught causing the REST API to except.