Process on finish call without using LoopCommunicator
In PR #4154, the LoopCommunicator was used as the runner's communicator.
This makes it easy to schecule callbacks into the correct event loops,
however, the `Process` has its own mechanism to do this. In object
method `_schedule_rpc` process schedule the callback into its event
loop. Using `LoopCommunitor` lead to a problem that when controller send
the controlling message by rpc for example when calling
`controller.pause_process` the `pause` subscriber is run by
`plumpy.futures.create_task`, which is not necessary.

`call_on_process_finish` receive a callback which not expected to be a
coroutine, however, in `action_awaitables` we schedule `_run_task` which
is a corotinefunction to it. The `call_soon` method of Process object is
more proper here.