`CalcJob`: move job resource validation to the `Scheduler` class
This is necessary because the resource validation is scheduler
dependent. Up till now, the `CalcJob` defined a validator on the entire
input namespace that validates the `metadata.options.resources` input.
Specifically, it demanded that the `num_machines` keyword was defined,
which is indeed a requirement for the `NodeNumberJobResource`, used for
schedulers like SLURM, however, this field doesn't even make sense for
schedulers like LSF and SGE that use the `ParEnvJobResource` type of job
resources.

The solution is to delegate the validation of the resources to the
scheduler which in turn delegates it to the `JobResource` class that it
uses. The validation used to happend on construction of the job resource
instance, but is factored out to the `validate_resources` classmethod.
This allows it to be called without having to construct an instance
which is more efficient.

Finally, the signature of the `CalcJob` input validators is changed and
no longer raise an `InputValidationError` but instead return an error
message as the interface of `Port.validator` requires. The port itself
will detect if an error message is returned and raise a `ValueError`
with all the relevant error messages.