Refactor the scheduler base class and data structure classes
The `Scheduler.get_valid_schedulers` method is deprecated as the
`aiida.plugins` module is reserved to inquire about installed plugins.
Furthermore, it is properly marked as an abstract class which allows to
remove the explicit `raise NotImplementedError` from methods that are
abstract and have to be implemented by subclasses. Finally, there is
some minor cleaning up of the code and docstrings.

The implementation of the `validate_resources` method of the subclasses
of `JobResource` has been improved and significantly simplified in the
case of `NodeNumberJobResource` making the logic a lot more readable.
Tests are added for the `ParEnvJobResource` job resource class which was
completely untested.