`CalcJob`: only use MPI procs per machine if scheduler supports it
The `CalcJob.presubmit` method was always adding the resource keyword
`num_mpiprocs_per_machine` if it was not already (implicitly) specified
and the computer defined a default. However, this keyword is not valid
for all schedulers, and for example, the SGE and LSF scheduler will
raise if it is passed. Here we add the `Scheduler.preprocess_resources`
method, which can be called and will properly take the type of job
resources into account when deciding whether `num_mpiprocs_per_machine`
should be defined.