Add transfer calcjob
This calcjob allows the user to copy files between a remote machine and
the local machine running AiiDA. More specifically, it can do any of the
following:

* Take any number of files from any number of RemoteData folders in
a remote machine and copy them in the local repository of a single
newly created FolderData node.

* Take any number of files from any number of FolderData nodes in the
local machine and copy them in a single newly created RemoteData folder
in a given remote machine.

These are the main two use cases, but there are also other more complex
combinations allowed by the current implementation.
Please check the documentation for more details.