Docs: update ReadTheDocs configuration file
This forces RTD to install the `aiida-core` package such that the entry
points are installed and it will now fail the build if warnings are
emitted.

Also remove the changes of the previous commit to manually call reentry
scan and the `ON_RTD` variable, which was a remnant of the old
documentation build.

Finally, delete folder `aiida/sphinxext/tests` which became obsolete
when all tests were moved to the `tests` folder.