Docs: update ReadTheDocs configuration and build process (#4203)
The `.readthedocs.yml` configuration file is updated to ask RTD to also
install the `aiida-core` package. This is necessary for the entry points
to be properly installed, which is turn is necessary for the automated
API documentation to be successfully built. This was always failing but
went unnoticed because the nitpick option was not enabled. This is also
changed such that failures like these will fail the build from now on.

Due to the package now being installed through `pip`  it is no longer
necessary, nor preferable, to hack it into the `sys.path` in the source
configuration file. It also allows to remote the change of the previous
commit that manualy called `reentry scan`, since this was not the actual
reason for the entry points not being available.

Since the package is now installed directly through `pip` the separate
`docs/requirements_for_rtd.txt` file is now obsolete and it is removed
along with all the pre-commit hooks and scripts that automatically
updated it.

Finally, delete folder `aiida/sphinxext/tests` which became obsolete
when all tests were moved to the `tests` folder.