Implement `skip_orm` option for SqlAlchemy `Group.remove_nodes`
The current implementation of `Group.remove_nodes` is very slow. For a
group of a few tens of thousands of nodes, removing a thousand can take
more than a day. The same problem exists for `add_nodes` which is why a
shortcut was added to the backend implementation for SqlAlchemy. Here,
we do the same for `remove_nodes`. The `SqlaGroup.remove_nodes` now
accepts a keyword argument `skip_orm` that, when True, will delete the
nodes by directly constructing a delete query on the join table.