Un-vendor the `fastentrypoints` dependency.
The `fastentrypoints` dependency is used to make console scripts
faster, in the case where they are **not** installed as wheels.
If they are installed as wheels, the console scripts are faster
by default.
By specifying `fastentrypoints` as a build requirement in
`pyproject.toml`, we can ensure that it can be imported whenever
the package is installed through `pip`.
This incurs a slight overhead of having to install
`fastentrypoints` even when the package is installed as a wheel,
but we no longer have to vendor `fastentrypoints`.