`BaseRestartWorkChain`: do not assume `metadata` exists in inputs
The contract of the base restart work chain is that the `setup` will
define the `self.ctx.inputs` attribute, however, it is not defined what
its contents should be. The `run_process` method was unjustly assuming
that the `metadata` key would be set to a dictionary, which is not
always the case. In this fix, we simply check for its existence and
create the dictionary if it doesn't yet exist.