Remove all files from the pre-commit exclude list (#4196)
Except for the documentation of course, which should remain excluded.

Also move the `pylint` pre-commit hook back under `local`. The idea by
putting it under the remote repos was to profit from the separate
virtual environment that is created with the exact version specified in
order to prevent clashes with requirements of other projects being
developed in the same virtual environment. However, this approach leads
to many spurious false positive import-errors because `pylint` cannot
find all other third party dependencies. This can be fixed by specifying
`language: system`, but this just forces the normal virtual environment
to be used, rendering the whole point of using the remote repos moot.