rename test fixtures => static
The name "fixtures" is currently used both for the pytest fixtures and
for test data in tests/fixtures, such as AiiDA export files.
This is confusing and makes searching in the codebase unnecessarily
difficult.
Here, we rename the test data folder to "static", which indicates the
static nature of the files residing there, while avoiding a clash of
definition with the pytest fixures residing in aiida.manage.tests.