add --in-place flag to verdi export migrate
When export migration is required, you can often get rid of the outdated
export file.
The `--in-place` flag saves users from specifying a temporary filename,
removing the old file and renaming the temporary one.
Using a temporary directory avoids the use of protected tempfile
functions and guarantees that there are no filename clashes.

This commit also changes the exit code for migrating an export file that
already is up to date from an error to success (0).

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>
Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>