CI: run all jobs on Python 3.8 (#4229)
The `pre-commit` and `verdi` steps of the CI workflow started failing.
Weirdly enough, it only seems to fail for the `aiidateam` repository but
the same branches on some other forks run fine. It might be due to the
Python version being used not being compatible with the runner, so we
update it to 3.8 for now.