Docs: revise the how-to write plugin for external codes (#4208)
The how-to section on how to create a calculation job and parser plugin
is significantly changed, in order to make it more self-contained. The
new version really attempts to give a complete manual from start to
finish to constructing new plugins and running them. This therefore
includes creating a minimal package with entry points, which is necessary
for the `Parser` to be able to be specified as the parser for the
calculation job. It also includes a launch script.

The `ArithmeticAddParser` was copied to provide a simpler version that
does not do any error checking and returning of exit codes. This is done
such that the first time the parser interface is explained, there is as
little detail as possible. The code is included in the existing parser
module such that it can be tested and the snippet is literally included
to prevent it from going out of sync with the actual code.