CI: do not fail the build when the coverage upload fails (#4239)
This happens quite a bit and will fail the entire build. Since GHA does
not yet allow rerunning a single failed job, we have to restart the
entire build. Since on top of that we are currently just using the
coverage as an indicator, it seems excessive to fail the entire build if
the coverage upload of a single job fails.