Add the `verdi profile caching` command
This command will try to load the caching configuration from the caching
configuration file, for any given profile. This is useful for users to
make sure that the file is correctly named and contains correct syntax,
before actually starting to run new calculations.

To test this, a new pytest fixture is created that creates a completely
new and independent configuration folder, along with fixtures to create
profiles to add to the config and caching configuration files. Similar
code already exists for normal unittests, but this can be removed once
those have been refactored to pytests.