Docs: move in the documentation on caching (#4228)
Split off the technical details to a Topics section. It is placed under
the "Provenance" topic, as I don't think it needs its own top level
section. Even though caching only applies to calculation jobs at the
moment, and so one could argue to place it there, really this is a
current implementation detail and this is explained in the limitations.
Since it has fundamentally to do with the provenance, this seems like
the best fit.

Finally, an `important` block at the beginning of the how-to to explain
why it is not enabled by default and to warn users of the caveats,
which links to the topics section.