Fix pre-commit configuration to reinstate `pylint` running (#4258)
The `pylint` hook was ignored because the `hook` key in the `local`
category was defined twice. In this case `pre-commit` won't complain
that the file is invalid but the second just overwrites the first
declaration. In addition, the hook was missing the keys `entry`, which
is required, and the `types` key which restricts it to running only on
Python file.