Docs: move "Running on supercomputers" section to "Run codes" (#4242)
It is important for users to be aware of these mechanisms and the
potential problems it can cause if not respected, before they run into
problems. That is why it is better to move these instructions as close
as possible as to where most users will learn how to connect to remote
clusters.