`ArithmeticAddParser`: attach output before checking for negative value (#4267)
For the recent documentation revamp, the `ArithmeticAddCalculation` and
`ArithmeticAddParser` were simplified significantly, by getting rid off
as much as the unnecessary code, because they are being literally
included as example for the basic how-to create a code plugin.

A part that was removed was the `settings` input node that allowed to
change the behavior of the parser and allow negative sums instead of it
returning an exit code. This, however, in turn cause the Reverse Polish
Notation tests on Jenkins to fail. Since these tests are not required to
pass, the changes were merged without a fix.

In the scope of the RPN tests, negative sums are fine, which anyway is
just a mechanism to introduce some kind of failure mode for demonstration
purposes. To fix this, without making the logic of the parser more
complex, we simply change the order of attaching the output node and
performing the final check. Since the RPN workchains only check if the
output node is there and do not care about the exit status of the
calculation, they will happily continue and the code of the parser keeps
the same complexity.