CI: Update `setup-python` action to v2 in order to pin Python to 3.7.7 (#4265)
Python 3.7.8, which was installed by default has some issues with our
requirement for `pyyaml==5.1.2. Since we cannot release this requirement
very easily, we temporarily workaround it by pinning 3.7.7 for the tests
job, which is the only job requiring Python 3.7. Note that this required
`setup-python@v2` since v1 does not allow specifying an exact version.