Make the loglevel of the daemonizer configurable (#4276)
The loglevel of the daemonizer, currently `circus`, was hardcoded in the
code of the daemon client `aiida.engine.daemon.client.DaemonClient`,
even though the configuration options already provided a way to change
the logging level for the `circus` logger. The hardcoded value is now
replaced by fetching the value from the profile configuration.