`Transport`: add option to not use a login shell for all commands (#4271)
Both the `LocalTransport` as well as the `SshTransport` were using a
bash login shell, i.e., using the `-l` flag in the bash commands, in
order to properly load the user environment which may contain crucial
environment variables to be set or modules to be loaded.

However, for certain machines, the login shell will produce spurious
output that prevents AiiDA from properly parsing the output from the
commands that are executed. The recommended approach is to remove the
code that is producing the output, but this is not always within the
control of the user. That is why a `use_login_shell` option is added to
the `Transport` class that switches the use of a login shell.

The new option is added to the `verdi computer configure` command and as
such is stored in the `AuthInfo`. The new logic affects all bash commands
that are executed, including the `gotocomputer` command that follows a
slightly different code path.