Improve the storage effiency of the `repository_metadata` format
Since now all nodes will have a full representation of the virtual file
hierarchy of their repository content, which is stored in the database
in the `repository_metadata` column, it is crucial that the format of
that metadata is optimized to use as little data as possible, to prevent
the database from bloating unnecessarily.

Currently, we only define two file types:

 * Directories: which can contain an arbitrary number of sub objects
 * Files: which cannot contain objects, but instead have a unique hashkey

This means that we can serialize the metadata, without encoding the
explicit type. If it contains a key, it is necessarily a file. If it is
not a file, it has to be a directory.

The serialization format therefore can make do with simple nested
dictionaries, where each entry either contains the key `k`, indicating
it is a file, or `o`, which can be a dictionary itself, containing the
objects contained within that directory. Note that we can even safely
omit the `o` key for an empty dictionary. Since there is no `k` we know
it has to be a directory anyway.

This logic is implemented in `FileObject.from_serialized` that can fully
reconstruct all instance properties by inferring them from the sparse
serialized data.