Deprecate getter and setter methods of `Computer` properties (#4252)
From `aiida-core==1.0.0`, we have started using properties for getters
and setters of the basic attributes of ORM entities. Rule of thumb here
is that if the attribute corresponds to a column on the underlying
database model, a property is used.

In addition, `label` is the preferred name for the third entity
identifier, alongside the ID and UUID. This is already the case for most
entities, except for `Computer` which is still using `name`. Here,
`name` is deprecated and replaced for `label`. The changes are not yet
propagated to the backend, which will be done once the deprecated
resources are fully removed. This is fine because the backend is not
part of the public API so doesn't have to go through a deprecation path.
The `name` keyword in `Computer.objects.get` is also deprecated and
replaced by `label`.