rerun flaky tests
Over time, we've accumulated a handful of tests that pass just fine most
of the time, but fail every now and then when the bits align with the
wrong star.
While the right thing to do would be a deep dive into astrology, there's
always so many other things to do!

As a mitigation strategy, this PR marks those tests as 'flaky' and uses
the pytest-rerunfailures plugin to automatically try re-running those
tests a few times before considering them failed.

Note that this also makes it easier to identify flaky tests, a simple
`git grep mark.flaky` will do the trick.