Querybuilder init now accepts entity_type = ''
There was a bug when using the queryhelp of a previously defined
QueryBuilder which was appended a single query for the generic "Node"
class as a template in the initialization for a new query.

The underlying problem was that the append method was doing the wrong
checks to detect missing cls/entity_type.

Also modified a test to check for this specific case.