Add test fixtures that allow running tests only for specific db backend (#4279)
The new fixtures `skip_if_not_sqlalchemy` and `skip_if_not_django` are
added. Any test using this fixture will be skipped if the loaded profile
does not use the SqlAlchemy or Django backend, respectively. This allows
to move some tests that were defined in `tests/backends/aiida_sqlalchemy`
to be moved to the backend agnostic location.

Typically, in order to skip tests conditionally, the specific decorator
`pytest.mark.skipif` is used. However, these are executed at startup
time, at which point the test environment including the test profile has
not yet been loaded, making it impossible to determine what the database
backend is and therefore whether the test should be run. The fixture is
executed only after the the profile is loaded, making it possible to
know the backend.

A few other tests that in the backend specific folder were removed since
they were already tested in the corresponding `tests/orm/implementation`
files.