process function run in global runner
The runner loop designs is changed for process function, previously the
new runner is create with a new event loop to run the process function
(i.e workfunction and calcfunction). Now, the loop is create along with
 `Runner` object by calling `asyncio.get_event_loop()`. that is to say,
the event loop is get from the current thread, is the process function
is nested in another one, they share the same event loop.