`QueryBuilder`: Accept empty string for `entity_type` in `append` method (#4299)
The `append` method allows the entity that should be appended to be
defined by a class, through the `cls` argument, or as a string, through
the `entity_type` argument. The logic that validates that at least one
and at most one of these two arguments was defined by the caller was
bugged, since it did not compare with explicit `None` but generic falsy
values. This lead to `entity_type=''` raising an exception, even though
this is a valid entity type string and corresponds to the base `Node`
class.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>