`QueryBuilder`: add new join keywords to join extended provenance
The current keywords `with_ancestors` and `with_descendants` are the
only available join keywords to match sub graphs that span more than one
link. However, they intentionally only follow the data-provenance,
meaning only `CREATE` and `INPUT_CALC` links are followed. Often,
though, one would like to also include the logical provenance in this
search. This was intentionally omitted since this part of the graph can
include cycles due to the nature of the `RETURN` link, which in turn can
lead to infinite recursions when traversing the graph.

Here, the `with_extended_ancestors` and `with_extended_descendants`
keywords are added, that in addition to the data-provenance will also
follow `INPUT_WORK`, `CALL_CALC` and `CALL_WORK` links. This subset is
guaranteed to have no cycles but still provides more functionality in
querying the provenance graph.