`CalcJob`: improve scheduler resource validation error message
The `NodeNumberJobResource.validate_resources` contained a bug, such
that when one of the default fields was set to `None` it would not be
caught. Instead it would bubble up and be caught in the validator of the
`CalcJob` input namespace `validate_calc_job`, which did catch the
`TypeError` thrown by `Scheduler.validate_resources`.

The `TypeError` was thrown by `None` being cast to `int` in the
`NodeNumberJobResource.validate_resources`, however, at this stage, a
`None` value should be accepted. Only if an actual value is specified
should it be checked to be a valid integer. The check that sufficient
fields have none `None` values is done later on. This will provide a
more intuitive error message, instead of the vague message:

  'field must be greater than or equal to one'

that was raised due to the bug even when the field was not explicitly
set by the user, causing even more confusion.

Note that since the bug in the validate resources is now fixed, and
therefore properly raises `ValueError` in case of a problem, the input
validator no longer needs nor should catch the `TypeError`.

Finally, the `__str__` method is implemented for the `Scheduler` class
such that it formats in a nicer way in error messages.